inventory_retriever/vector/embedding_service.py
"""
Embedding Service for Warehouse Operations

Provides text embedding capabilities for semantic search over
warehouse documentation and operational procedures.
"""

import logging
from typing import List, Optional, Dict, Any
import numpy as np
import os
from dotenv import load_dotenv

load_dotenv()

logger = logging.getLogger(__name__)

class EmbeddingService:
    """
    Embedding service for generating vector representations of text.
    
    This is a placeholder implementation that can be extended to use
    actual embedding models like OpenAI, HuggingFace, or NVIDIA NIM.
    """
    
    def __init__(self, model_name: str = "default", dimension: int = 768):
        self.model_name = model_name
        self.dimension = dimension
        self._initialized = False
    
    async def initialize(self) -> None:
        """Initialize the embedding service."""
        try:
            # TODO: Initialize actual embedding model
            # For now, this is a placeholder
            self._initialized = True
            logger.info(f"Embedding service initialized with model: {self.model_name}")
        except Exception as e:
            logger.error(f"Failed to initialize embedding service: {e}")
            raise
    
    async def generate_embedding(self, text: str) -> List[float]:
        """
        Generate embedding for a single text.
        
        Args:
            text: Input text to embed
            
        Returns:
            List of float values representing the embedding
        """
        try:
            if not self._initialized:
                await self.initialize()
            
            # TODO: Replace with actual embedding generation
            # For now, return a random vector for testing
            np.random.seed(hash(text) % 2**32)
            embedding = np.random.normal(0, 1, self.dimension).tolist()
            
            logger.debug(f"Generated embedding for text: {text[:50]}...")
            return embedding
            
        except Exception as e:
            logger.error(f"Failed to generate embedding: {e}")
            raise
    
    async def generate_embeddings(self, texts: List[str]) -> List[List[float]]:
        """
        Generate embeddings for multiple texts.
        
        Args:
            texts: List of input texts to embed
            
        Returns:
            List of embeddings
        """
        try:
            if not self._initialized:
                await self.initialize()
            
            embeddings = []
            for text in texts:
                embedding = await self.generate_embedding(text)
                embeddings.append(embedding)
            
            logger.info(f"Generated {len(embeddings)} embeddings")
            return embeddings
            
        except Exception as e:
            logger.error(f"Failed to generate embeddings: {e}")
            raise
    
    async def similarity(self, embedding1: List[float], embedding2: List[float]) -> float:
        """
        Calculate cosine similarity between two embeddings.
        
        Args:
            embedding1: First embedding vector
            embedding2: Second embedding vector
            
        Returns:
            Cosine similarity score between -1 and 1
        """
        try:
            # Convert to numpy arrays
            vec1 = np.array(embedding1)
            vec2 = np.array(embedding2)
            
            # Calculate cosine similarity
            dot_product = np.dot(vec1, vec2)
            norm1 = np.linalg.norm(vec1)
            norm2 = np.linalg.norm(vec2)
            
            if norm1 == 0 or norm2 == 0:
                return 0.0
            
            similarity = dot_product / (norm1 * norm2)
            return float(similarity)
            
        except Exception as e:
            logger.error(f"Failed to calculate similarity: {e}")
            return 0.0
    
    def get_dimension(self) -> int:
        """Get the embedding dimension."""
        return self.dimension
    
    def get_model_name(self) -> str:
        """Get the model name."""
        return self.model_name

# Global embedding service instance
_embedding_service: Optional[EmbeddingService] = None

async def get_embedding_service() -> EmbeddingService:
    """Get or create the global embedding service instance."""
    global _embedding_service
    if _embedding_service is None:
        _embedding_service = EmbeddingService()
        await _embedding_service.initialize()
    return _embedding_service
