requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.7
httpx>=0.27
python-dotenv>=1.0
loguru>=0.7
langgraph>=0.2.30
asyncpg>=0.29.0
pymilvus>=2.3.0
numpy>=1.24.0
langchain-core>=0.1.0
aiohttp>=3.8.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
PyYAML>=6.0
prometheus-client>=0.19.0
httpx>=0.27.0
paho-mqtt>=1.6.0
websockets>=11.0.0
pymodbus>=3.0.0
bacpypes3>=0.0.0
# ERP Integration
requests>=2.31.0
# RFID/Barcode Scanning
pyserial>=3.5
# Time Attendance
pybluez>=0.23
